✨ Feat: 가게의 리뷰 목록 조회 api 구현
week9 - 실습

+++ src/main/java/umc/spring/dto/ReviewDTO/ReviewRequestDTO.java
@@ -16,7 +16,7 @@ public class ReviewRequestDTO {
         String description;
         @Min(1) @Max(5)
         @NotNull
-        Integer stars;
+        Float stars;
         @ExistMembers
         Long member;
         @ExistStores
